+++ CHANGELOG.rst
@@ -1,9 +1,10 @@
-v0.0.6 (November XX, 2013)
+v0.0.6 (February XX, 2014)
 --------------------------
 
   - ``environment`` parameter in the ``process`` module changed to ``env``.
     Creation of full environment moved to the base ``Process`` class.
   - Adding ``working_dir`` parameter to the ``Process`` class constructor.
+  - Rename tea.shutil to tea.shell
 
 
 v0.0.5 (October 12, 2013)